Extract app store version submission actions into separate action group
