+++ src/codemagic/tools/_app_store_connect/action_groups/__init__.py
@@ -1,3 +1,4 @@
+from .app_store_version_submissions_action_group import AppStoreVersionSubmissionsActionGroup
 from .apps_action_group import AppsActionGroup
 from .beta_app_review_submissions_action_group import BetaAppReviewSubmissionsActionGroup
 from .beta_build_localizations_action_group import BetaBuildLocalizationsActionGroup
